feat: add fee signature modal on withdrawal
